Change: Make the config settings evaluation more flexible
Allow to apply different kind of settings after each other. Hopefully
this will allow to read the rsync URL from the enterprise key and make
it visible in --help at the end.
